refactor: Change folder structure
- move atoms components into common
- rename BottomSheet component
